requirements.txt
Pillow
scikit-learn
pandas
seaborn
numpy
matplotlib
streamlit
streamlit-option-menu
tensorflow
gdown
